removed multiple 'expects' from tests. renamed test files to describe the components they test
