+++ src/context/session_context/session_context.tsx
@@ -10,7 +10,7 @@ import { AuthService } from '../../services/authentication_service/authenticatio
 import { UserProfile } from '../../schemas/user_profile_schema';
 import { UserToken } from '../../schemas/user_token_schema';
 import Axios, { AxiosInstance } from 'axios';
-import { ServiceProviderContext } from '../service_provider/service_provider_context';
+import { ServiceProviderContext } from '../service_provider_context/service_provider_context';
 
 export type AuthenticationState =
   | 'initial'
